split out outline checks
