+++ Lib/fontbakery/checks/outline_settings.py
@@ -1,0 +1,7 @@
+ALIGNMENT_MISS_EPSILON = 2  # Two point lee-way on alignment misses
+COLINEAR_EPSILON = 0.1  # Radians
+FALSE_POSITIVE_CUTOFF = 100  # More than this and we don't make a report
+JAG_ANGLE = 0.25  # Radians
+JAG_AREA_EPSILON = 0.05  # <5% of total outline area makes a jaggy segment
+SHORT_PATH_ABSOLUTE_EPSILON = 3  # 3 units is a small outline
+SHORT_PATH_EPSILON = 0.006  # <0.6% of total outline length makes a short segment
